listen for tx notifications on wallet opening
